File Clean and XyloSim added
Xylo.py contains the XyloSim, XyloSamna and XyloMonitor code. It detects if the hdk is connected and uses it if it is, otherwise it uses a simulation instead.

Doing channel wise quantisation.

Currently got random dummy data which I'm feeding the sim need to get some actual data samples instead.